refactor: move ZWaveOptions to separate file
