+++ packages/zwave-js/src/Driver.ts
@@ -1,6 +1,7 @@
 export { Driver } from "./lib/driver/Driver";
-export type { SendMessageOptions, ZWaveOptions } from "./lib/driver/Driver";
+export type { SendMessageOptions } from "./lib/driver/Driver";
 export type { FileSystem } from "./lib/driver/FileSystem";
+export type { ZWaveOptions } from "./lib/driver/ZWaveOptions";
 export {
 	FunctionType,
 	MessagePriority,
