+++ packages/zwave-js/src/lib/driver/Driver.ts
@@ -114,7 +114,6 @@ import {
 	compileStatistics,
 	sendStatistics,
 } from "../telemetry/statistics";
-import type { FileSystem } from "./FileSystem";
 import {
 	createSendThreadMachine,
 	SendThreadInterpreter,
@@ -123,6 +122,7 @@ import {
 } from "./SendThreadMachine";
 import { throttlePresets } from "./ThrottlePresets";
 import { Transaction } from "./Transaction";
+import type { ZWaveOptions } from "./ZWaveOptions";
 
 // eslint-disable-next-line
 const { version: libVersion } = require("../../../package.json");
@@ -135,95 +135,6 @@ const libNameString = `
 ███████╗ ╚███╔███╔╝ ██║  ██║  ╚████╔╝  ███████╗        ╚█████╔╝ ███████║
 ╚══════╝  ╚══╝╚══╝  ╚═╝  ╚═╝   ╚═══╝   ╚══════╝         ╚════╝  ╚══════╝
 `;
-
-export interface ZWaveOptions {
-	/** Specify timeouts in milliseconds */
-	timeouts: {
-		/** how long to wait for an ACK */
-		ack: number; // >=1, default: 1000 ms
-		/** not sure */
-		byte: number; // >=1, default: 150 ms
-		/**
-		 * How long to wait for a controller response. Usually this timeout should never elapse,
-		 * so this is merely a safeguard against the driver stalling
-		 */
-		response: number; // [500...5000], default: 1600 ms
-		/** How long to wait for a callback from the host for a SendData[Multicast]Request */
-		sendDataCallback: number; // >=10000, default: 65000 ms
-		/** How much time a node gets to process a request and send a response */
-		report: number; // [1000...40000], default: 10000 ms
-		/** How long generated nonces are valid */
-		nonce: number; // [3000...20000], default: 5000 ms
-
-		/**
-		 * @internal
-		 * How long to wait for a poll after setting a value
-		 */
-		refreshValue: number;
-	};
-
-	attempts: {
-		/** How often the driver should try communication with the controller before giving up */
-		controller: number; // [1...3], default: 3
-		/** How often the driver should try sending SendData commands before giving up */
-		sendData: number; // [1...5], default: 3
-		/** Whether a command should be retried when a node acknowledges the receipt but no response is received */
-		retryAfterTransmitReport: boolean; // default: false
-		/**
-		 * How many attempts should be made for each node interview before giving up
-		 */
-		nodeInterview: number; // [1...10], default: 5
-	};
-
-	/**
-	 * Optional log configuration
-	 */
-	logConfig?: LogConfig;
-
-	/**
-	 * @internal
-	 * Set this to true to skip the controller interview. Useful for testing purposes
-	 */
-	skipInterview?: boolean;
-
-	storage: {
-		/** Allows you to replace the default file system driver used to store and read the cache */
-		driver: FileSystem;
-		/** Allows you to specify a different cache directory */
-		cacheDir: string;
-		/**
-		 * How frequently the values and metadata should be written to the DB files. This is a compromise between data loss
-		 * in cause of a crash and disk wear:
-		 *
-		 * * `"fast"` immediately writes every change to disk
-		 * * `"slow"` writes at most every 5 minutes or after 500 changes - whichever happens first
-		 * * `"normal"` is a compromise between the two options
-		 */
-		throttle: "fast" | "normal" | "slow";
-	};
-
-	/** Specify the network key to use for encryption. This must be a Buffer of exactly 16 bytes. */
-	networkKey?: Buffer;
-
-	/**
-	 * Some Command Classes support reporting that a value is unknown.
-	 * When this flag is `false`, unknown values are exposed as `undefined`.
-	 * When it is `true`, unknown values are exposed as the literal string "unknown" (even if the value is normally numeric).
-	 * Default: `false` */
-	preserveUnknownValues?: boolean;
-
-	/**
-	 * Some SET-type commands optimistically update the current value to match the target value
-	 * when the device acknowledges the command.
-	 *
-	 * While this generally makes UIs feel more responsive, it is not necessary for devices which report their status
-	 * on their own and can lead to confusing behavior when dealing with slow devices like blinds.
-	 *
-	 * To disable the optimistic update, set this option to `true`.
-	 * Default: `false`
-	 */
-	disableOptimisticValueUpdate?: boolean;
-}
 
 const defaultOptions: ZWaveOptions = {
 	timeouts: {
